requirements.txt
python-telegram-bot==13.4.1
requests==2.25.1
dataclasses-json==0.5.2